@@ -8,10 +8,10 @@
   "publisher": "JosephAbbey",
   "displayName": "Scratch VHDL",
   "description": "",
-  "version": "0.2.12",
+  "version": "1.0.0",
   "license": "MIT",
   "engines": {
-    "vscode": "^1.65.0"
+    "vscode": "^1.75.0"
   },
   "categories": [
     "Other"
@@ -81,6 +81,12 @@
         ]
       }
     ],
+    "jsonValidation": [
+      {
+        "fileMatch": "*.sbe",
+        "url": "./sbe.schema.json"
+      }
+    ],
     "configuration": {
       "title": "ScratchVHDL",
       "properties": {
@@ -99,21 +105,30 @@
     }
   },
   "scripts": {
-    "vscode:prepublish": "npm run compile",
-    "compile": "tsc -p ./",
-    "lint": "eslint . --ext .ts,.tsx",
-    "watch": "tsc -w -p ./"
+    "vscode:prepublish": "pnpm package",
+    "compile": "node ./esbuild.cjs",
+    "package": "cross-env NODE_ENV=production node ./esbuild.cjs",
+    "watch": "node ./esbuild.cjs --watch",
+    "lint": "eslint src --ext ts",
+    "generate-schema": "node ./generate-schema.cjs"
   },
   "devDependencies": {
-    "@types/node": "^16.11.7",
-    "@types/vscode": "^1.65.0",
-    "@typescript-eslint/eslint-plugin": "^5.30.0",
-    "@typescript-eslint/parser": "^5.30.0",
-    "eslint": "^8.13.0",
+    "@types/glob": "^8.0.0",
+    "@types/node": "16.x",
+    "@types/vscode": "^1.75.0",
+    "@types/vscode-webview": "^1.57.0",
+    "@typescript-eslint/eslint-plugin": "^5.42.0",
+    "@typescript-eslint/parser": "^5.42.0",
+    "cross-env": "^7.0.3",
+    "esbuild": "^0.16.10",
+    "eslint": "^8.26.0",
+    "glob": "^8.0.3",
+    "prettier": "^2.8.1",
+    "ts-json-schema-generator": "^1.3.0",
     "typescript": "^4.8.4"
   },
   "dependencies": {
-    "@blockly/workspace-backpack": "^3.0.2",
-    "blockly": "^9.0.0"
+    "@vscode/webview-ui-toolkit": "^1.2.2",
+    "blockly": "^10.1.3"
   }
 }
